backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Task Queue
celery==5.3.6
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.6
cryptography==42.0.0

# AWS & Storage
boto3==1.34.34
botocore==1.34.34

# Video Processing
ffmpeg-python==0.2.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3
requests==2.31.0
requests-oauthlib==1.3.1

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0.post1
croniter==2.0.1
python-json-logger==2.0.7
python-magic==0.4.27

# Monitoring
sentry-sdk[fastapi]==1.40.0
prometheus-client==0.19.0

# Rate Limiting
slowapi==0.1.9

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
